Reorder structs and modules
