+++ plugin/src/process/audio/pair.rs
@@ -1,44 +1,61 @@
-use crate::prelude::Audio;
-use crate::process::audio::channels::ChannelPair::{InputOnly, OutputOnly};
-use crate::process::audio::channels::{ChannelPair, SampleType};
-use crate::process::audio::port::{AudioInputPort, AudioOutputPort};
+use crate::process::audio::pair::ChannelPair::{InputOnly, OutputOnly};
+use crate::process::audio::{InputPort, OutputPort, SampleType};
+use crate::process::Audio;
 use clap_sys::audio_buffer::clap_audio_buffer;
 use std::slice::{Iter, IterMut};
 
-pub struct AudioPortPair<'a> {
+pub struct PortPair<'a> {
     input: Option<&'a clap_audio_buffer>,
     output: Option<&'a mut clap_audio_buffer>,
     frames_count: u32,
 }
 
-impl<'a> AudioPortPair<'a> {
+impl<'a> PortPair<'a> {
     #[inline]
-    pub fn input(&self) -> Option<AudioInputPort<'a>> {
-        // SAFETY: TODO
+    pub fn input(&self) -> Option<InputPort<'a>> {
         self.input
-            .map(|i| unsafe { AudioInputPort::from_raw(i, self.frames_count) })
+            .map(|i| unsafe { InputPort::from_raw(i, self.frames_count) })
     }
 
     #[inline]
-    pub fn output(&mut self) -> Option<AudioOutputPort> {
-        // SAFETY: TODO
+    pub fn output(&mut self) -> Option<OutputPort> {
         self.output
             .as_mut()
-            .map(|i| unsafe { AudioOutputPort::from_raw(i, self.frames_count) })
+            .map(|i| unsafe { OutputPort::from_raw(i, self.frames_count) })
     }
 
-    fn input_channel(&self, index: usize) -> Option<SampleType<&[f32], &[f64]>> {
-        todo!()
+    #[inline]
+    fn input_channel(&self, index: usize) -> Option<SampleType<&'a [f32], &'a [f64]>> {
+        Some(
+            unsafe { SampleType::from_raw_buffer(self.input?) }?
+                .map_option(|b| b.get(index), |b| b.get(index))?
+                .map(
+                    |b| unsafe { core::slice::from_raw_parts(*b, self.frames_count as usize) },
+                    |b| unsafe { core::slice::from_raw_parts(*b, self.frames_count as usize) },
+                ),
+        )
     }
 
-    fn output_channel(&self, index: usize) -> Option<SampleType<&mut [f32], &mut [f64]>> {
-        todo!()
+    #[inline]
+    fn output_channel(&mut self, index: usize) -> Option<SampleType<&'a mut [f32], &'a mut [f64]>> {
+        Some(
+            unsafe { SampleType::from_raw_buffer_mut(self.output.as_mut()?) }?
+                .map_option(|b| b.get(index), |b| b.get(index))?
+                .map(
+                    |b| unsafe {
+                        core::slice::from_raw_parts_mut(*b as *mut _, self.frames_count as usize)
+                    },
+                    |b| unsafe {
+                        core::slice::from_raw_parts_mut(*b as *mut _, self.frames_count as usize)
+                    },
+                ),
+        )
     }
 
     pub fn channel_pair(
         &mut self,
         index: usize,
-    ) -> Option<SampleType<ChannelPair<f32>, ChannelPair<f64>>> {
+    ) -> Option<SampleType<ChannelPair<'a, f32>, ChannelPair<'a, f64>>> {
         let input = self.input_channel(index);
         let output = self.output_channel(index);
 
@@ -47,7 +64,7 @@ impl<'a> AudioPortPair<'a> {
 
     pub fn channel_pairs(
         &mut self,
-    ) -> Option<SampleType<AudioChannelPairsIter<'a, f32>, AudioChannelPairsIter<'a, f64>>> {
+    ) -> Option<SampleType<ChannelPairsIter<'a, f32>, ChannelPairsIter<'a, f64>>> {
         let input = self
             .input
             .and_then(|b| unsafe { SampleType::from_raw_buffer(b) })
@@ -67,12 +84,12 @@ impl<'a> AudioPortPair<'a> {
 
         input.try_match_with(output).map(|s| {
             s.map(
-                |(i, o)| AudioChannelPairsIter {
+                |(i, o)| ChannelPairsIter {
                     input_iter: i,
                     output_iter: o,
                     frames_count: self.frames_count,
                 },
-                |(i, o)| AudioChannelPairsIter {
+                |(i, o)| ChannelPairsIter {
                     input_iter: i,
                     output_iter: o,
                     frames_count: self.frames_count,
@@ -82,13 +99,13 @@ impl<'a> AudioPortPair<'a> {
     }
 }
 
-pub struct AudioChannelPairsIter<'a, S> {
+pub struct ChannelPairsIter<'a, S> {
     input_iter: Iter<'a, *const S>,
     output_iter: IterMut<'a, *const S>,
     frames_count: u32,
 }
 
-impl<'a, S> Iterator for AudioChannelPairsIter<'a, S> {
+impl<'a, S> Iterator for ChannelPairsIter<'a, S> {
     type Item = ChannelPair<'a, S>;
 
     #[inline]
@@ -137,13 +154,13 @@ fn channel_pair_from_io<'a>(
     }
 }
 
-pub struct AudioPortsPairIter<'a> {
+pub struct PortsPairIter<'a> {
     inputs: Iter<'a, clap_audio_buffer>,
     outputs: IterMut<'a, clap_audio_buffer>,
     frames_count: u32,
 }
 
-impl<'a> AudioPortsPairIter<'a> {
+impl<'a> PortsPairIter<'a> {
     #[inline]
     pub(crate) fn new(audio: &'a mut Audio<'_>) -> Self {
         Self {
@@ -154,17 +171,35 @@ impl<'a> AudioPortsPairIter<'a> {
     }
 }
 
-impl<'a> Iterator for AudioPortsPairIter<'a> {
-    type Item = AudioPortPair<'a>;
+impl<'a> Iterator for PortsPairIter<'a> {
+    type Item = PortPair<'a>;
 
     fn next(&mut self) -> Option<Self::Item> {
         match (self.inputs.next(), self.outputs.next()) {
             (None, None) => None,
-            (input, output) => Some(AudioPortPair {
+            (input, output) => Some(PortPair {
                 input,
                 output,
                 frames_count: self.frames_count,
             }),
+        }
+    }
+}
+
+pub enum ChannelPair<'a, S> {
+    InputOnly(&'a [S]),
+    OutputOnly(&'a mut [S]),
+    InputOutput(&'a [S], &'a mut [S]),
+    InPlace(&'a mut [S]),
+}
+
+impl<'a, S> ChannelPair<'a, S> {
+    #[inline]
+    pub(crate) fn from_io(input: &'a [S], output: &'a mut [S]) -> Self {
+        if input.as_ptr() == output.as_ptr() {
+            ChannelPair::InPlace(output)
+        } else {
+            ChannelPair::InputOutput(input, output)
         }
     }
 }
